pyproject.toml
[project]
name = "examen-dvc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.67.0",
    "dvc-s3>=3.3.0",
    "pandas>=3.0.1",
    "pyaml>=26.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
]
